refactor(services): audit_service 를 패키지로 승격해 한도 안에 넣는다 (B5.5 Task 1)
파일 분할 프로그램 B5.5 의 첫 대상. 986 줄 → 최대 모듈 633 줄.

착수 전 실측이 계획서 분류를 뒤집었다
계획서(2026-08-09)는 B5.5 5 개를 "집중도 48~65% → 2 단계(클래스 이동 + 메서드
추출) 필요" 로 분류했다. 집중도(최대 클래스 / 파일)는 판정 지표가 아니다.
실제 지표는 잔여 = 전체 − 이동가능이고, 그 값으로 보면 5 개 중 4 개가 정의
이동만으로 한도에 들어간다:

| 대상 | 줄수 | 집중도 | 잔여 | 문자열패치 | 모듈객체패치 |
|---|---:|---:|---:|---:|---:|
| audit_service | 986 | 56% | 634 | 0 | 0 |
| tmux_service | 921 | 63% | 660 | 4 | 0 |
| merge_service | 1,329 | 48% | 693 | 6 | 0 |
| notification_service | 1,018 | 65% | 743 | 13 | 0 |
| playground_service | 1,287 | 63% | 963 | 0 | 12 |

첫 대상을 고른 근거는 잔여가 아니라 패치 스캔이다 — audit_service 만 네 형태
전부 0 건이다. tmux_service 의 4 건은 모듈 지역 타깃이라 계획서가 orchestrator.
nodes 에서 "비관대" 로 분류한 형태이고, playground_service 는 2 단계인 데다
모듈 객체 패치 12 건 + 직접 참조까지 있어 마지막이다.

배치
패키지 승격(audit_service.py → audit_service/__init__.py)으로 20 여 개 import
사이트에서 services.audit_service 가 글자 그대로 유효하게 남는다. B4 가 확정한
"메인 파일을 옮기지 않는다" 의 파이썬 형태다.

- service.py 633 — AuditService · _audit_logs · _filter_logs · USE_DATABASE
- models.py  155 — 열거형·Pydantic 5 종. 서로만 참조하는 잎
- helpers.py 151 — audit_* 래퍼 7 종
- stats.py   101 — 집계 3 종 + 카테고리 상수. 저장소를 건드리지 않는다
- __init__.py 55 — 배럴 (외부가 가져가는 15 종 재노출)

순수 이동이 아닌 곳이 정확히 한 군데 있다 — cleanup_old_logs
분할이 회귀를 만들었고 Codex 리뷰가 잡았다. 실측으로 확인한 동작 차이:

    분할 전  import services.audit_service as m; m._audit_logs  -> purge 반영(1건)
    분할 후  배럴 재노출                                        -> 낡음(2건)

원인은 cleanup_old_logs 의 global 재바인딩이다. global 은 정의한 모듈의 이름만
다시 묶으므로, 단일 모듈일 때는 그 모듈 속성이 곧 최신이었지만 배럴의
from .service import _audit_logs 는 스냅샷이라 purge 이전 리스트를 계속 노출한다.

고친 방향은 배럴을 동적으로 만드는 것이 아니라 재바인딩 자체를 없애는 것이다 —
_audit_logs[:] = [...] 로 제자리 변경한다. 모든 보유자가 같은 내용을 보게 되고,
"저장소·_filter_logs·AuditService 를 같은 모듈에 둬야 한다" 는 제약도 함께
무의미해진다. AuditService 본문 변경은 이 메서드 한 곳(13 줄)뿐이며 감사 도구가
그것을 정확히 지목한다.

여기서 한 번 과하게 주장했다가 되돌렸다
초안은 "저장소를 떼면 깨진다" 고 단언했으나 증명하지 못했다. 금지 배치를 만들면
두 모듈 전역이 갈라지는 것은 맞지만 그 차이를 읽는 소비자가 0 건이었고, 그 증거가
초안 테스트다 — 금지 배치에서 16/16 통과했다. 그리고 그 판단에 기대어 "배럴이
낡는 것은 기존 동작" 이라고 넘긴 것이 위 회귀를 놓친 경로다. 소비자가 없다는
사실이 동작이 안 바뀌었다는 근거가 되지는 않는다.

검증
- 정의 보존: scripts/split_audit.py 신규. 라인 범위로 잘라 그 diff 를 보면
  자기확인이 되므로 AST 이름 집합 + 본문 바이트 대조로 검사한다.
  22 종 → 22 종, 유실·신규 0, 본문 변경 1 종(AuditService, 위 의도된 변경)
- 도구 자체를 Red-Green 했다: 원본 복사 OK / 정의 1 종 삭제 FAIL /
  본문 한 구절 변경 FAIL / 데코레이터만 제거 FAIL /
  if TYPE_CHECKING 안의 정의 삭제 FAIL / try 분기 정의 삭제 FAIL /
  for·while·match 안의 정의 삭제 FAIL
- 배럴 표면: 외부가 가져가는 15 종을 import 사이트에서 실측해 테스트로 고정.
  하나라도 빠지면 소비자가 import 시점에 죽는데 패키지 단위 테스트로는 안 잡힌다
- pytest 1685 passed, 33 skipped, 0 failed (기준선 1668/33 + 신규 17)
- ruff check / ruff format --check / mypy 317 files 0 issues

Codex 리뷰가 검증 도구의 결함 3 건도 잡았고 전부 고쳤다
도구에 무판정 구멍이 있으면 분할이 통과한 것도 근거가 되지 않으므로 각각 RED 로
재현한 뒤 고쳤다:
- 문서화한 호출법(audit_service/*.py)이 배럴의 __all__ 을 "신규 정의" 로 세어
  올바른 분할을 FAIL 로 보고했다. 울리는 게이트는 결국 배럴을 빼고 돌리게 되고
  그러면 배럴을 아예 안 보게 된다
- 데코레이터가 node.lineno 위에 있어 비교 본문에서 빠졌다. 데코레이터만 지워도
  "변경 없음" 이었다
- tree.body 만 봐서 if TYPE_CHECKING · try 안의 정의를 무시했다. 이어진 라운드가
  for · while · match 도 같은 구멍임을 짚어 함께 덮었다 — 검증기가 suite 하나를
  건너뛰면 그 안에서 사라진 정의에 대해 성공을 보고한다
재귀를 넣자 try/except 재대입이 "중복 정의" 로 죽어, 파일 안의 재대입은 순서대로
이어 붙이고 중복 판정은 모듈 간에만 적용하도록 바꿨다.

반영하지 않은 지적 1 건과 그 이유
"import 도 바인딩이니 감사에 포함하라" 는 위험 자체는 맞다. 그러나 분할은 같은
이름을 여러 모듈에서 다시 import 하므로 1:1 대응을 요구하면 올바른 분할이 전부
실패한다(실제로 AuditAction 이 배럴과 helpers 에서 충돌했다). 이 층은 도구가
아니라 ruff F821 이 정적으로 덮는다 — 주장이 아니라 실측이다: helpers.py 에서
import 한 줄을 지우면 ruff 가 F821 Undefined name `AuditService` 를 낸다.
도구 docstring 에 그 경계를 적었다.
